SRCAndMore/pyproject.toml
[project]
name = "mady-bot"
version = "6.2.1"
description = "Mady Bot (Codename: Heather) - Telegram Payment Gateway Testing Bot"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "beautifulsoup4>=4.14.3",
    "cloudscraper>=1.2.71",
    "cryptography>=46.0.3",
    "faker>=40.1.0",
    "lxml>=6.0.2",
    "playwright>=1.57.0",
    "psycopg2-binary>=2.9.11",
    "pycryptodome>=3.23.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "telegram>=0.0.1",
    "urllib3>=2.6.3",
    "user-agent>=0.1.14",
]
